refactore search files
