+++ search/__init__.py
@@ -1,0 +1,2 @@
+from .puzzle_env import SlidingPuzzleEnv
+from .search import Search
